refactor: extract GenerateDialog
